update homework script

+++ scripts/homework.sh
@@ -1,23 +1,42 @@
 #!/bin/bash
 
-dirName="322_submission_`whoami`" ;
+dirName="322_submission" ;
 origDir=`pwd` ;
 
 # Copy all files to an empty directory
+if test -d ../$dirName ; then
+  echo "ERROR: Please delete ../$dirName" ;
+  exit 1 ;
+fi
 mkdir ../$dirName ;
 cd ../$dirName ;
-cp -r $origDir/* ./ ;
+cp -r "${origDir}"/* ./ ;
 
 # Remove unnecessary files
 rm -r lib Makefile scripts bin ;
 rm -r */tests ;
-rm -r */obj ;
-rm -r */bin ;
+rm -fr */obj ;
+rm -fr */bin ;
 rm -r */scripts ;
 rm  */Makefile ;
+for i in `ls` ; do
+  if ! test -d $i ; then
+    continue ;
+  fi
+  pushd ./ ;
+  cd $i ;
+  for j in `ls` ; do
+    if test -d $j ; then
+      continue ;
+    fi
+    rm -f $j ;
+  done
+  popd ;
+done
 
 # Create the package
 cd ../ ;
+rm -f ${dirName}.tar.bz2 ;
 tar cfj ${dirName}.tar.bz2 ${dirName} ;
-mv ${dirName}.tar.bz2 src ;
+mv ${dirName}.tar.bz2 "${origDir}"/ ;
 rm -r ${dirName} ;
